+++ biggygains/components/sentiment/interface.py
