+++ biggygains/trading/stock.py
